Move column_* files into parquet/column folder
